gis-emergency-python/backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
psycopg2-binary==2.9.7
geoalchemy2==0.14.2
geopy==2.4.0
requests==2.31.0
python-dotenv==1.0.0
shapely==2.0.1
pytest==7.4.2